Separated types and accessors to data
